+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Updated database fixtures
 * Updated AppVeyor configuration
 
 2.2.3